(feat) CoffeeScript support
Generator can create CoffeeScript application source code. Previous
generator functionality for JavaScript exists for CoffeeScript (base
code, adding modules, adding routes, etc.). Code coverage is not currently
supported for CoffeeScript application source code.

+++ route/index.js
@@ -14,10 +14,6 @@ Generator.prototype.prompting = function prompting() {
 Generator.prototype.writing = function writing() {
   var config = this.getConfig()
 
-    // load app.js to prepare adding new state
-    , filePath = path.join(this.config.path, '../app/', config.modulePath, utils.hyphenName(config.moduleName) + '.js')
-    , file = fs.readFileSync(filePath, 'utf8')
-
     ,  newState = {
       module: this.module,
       url: this.url,
@@ -32,9 +28,25 @@ Generator.prototype.writing = function writing() {
       controllerAs: config.controllerAs,
       passFunc: config.passFunc,
       ngRoute: config.ngRoute
-    };
+    }
 
-  fs.writeFileSync(filePath, utils.addRoute(file, newState, newRouteConfig));
+    // module file to add route to
+    , filePath, file;
+
+  filePath = path.join(this.config.path, '../app/', config.modulePath, utils.hyphenName(config.moduleName) + '.coffee');
+
+  // load JavaScript app if CoffeeScript app doesn't exist
+  if (fs.existsSync(filePath)) {
+    file = fs.readFileSync(filePath, 'utf8');
+
+    fs.writeFileSync(filePath, utils.addRouteCoffee(file, newState, newRouteConfig));
+  } else {
+    filePath = path.join(this.config.path, '../app/', config.modulePath, utils.hyphenName(config.moduleName) + '.js');
+
+    file = fs.readFileSync(filePath, 'utf8');
+
+    fs.writeFileSync(filePath, utils.addRoute(file, newState, newRouteConfig));
+  }
 
   // e2e testing
   // create page object model
